Deduplicate shared code between desktop apps into motorsports_data_notebook.desktop
Move loader.py, SessionPanel, get_screen_dpi(), and setup_hidpi_scaling()
from both desktop apps into the shared motorsports_data_notebook.desktop
package. SessionPanel gains an auto_select parameter ("best" or "top_103")
to handle the one behavioral difference between the two apps.
