+++ desktop_app/src/suspension_analyzer/widgets/__init__.py
@@ -1,7 +1,6 @@
 """UI widgets for Suspension Analyzer."""
 
-from suspension_analyzer.widgets.session_panel import SessionPanel
 from suspension_analyzer.widgets.config_panel import ConfigPanel
 from suspension_analyzer.widgets.chart_view import ChartView
 
-__all__ = ["SessionPanel", "ConfigPanel", "ChartView"]
+__all__ = ["ConfigPanel", "ChartView"]
