+++ desktop_app_driver/src/driver_consistency_analyzer/widgets/__init__.py
@@ -1,8 +1,7 @@
 """UI widgets for Driver Consistency Analyzer."""
 
-from driver_consistency_analyzer.widgets.session_panel import SessionPanel
 from driver_consistency_analyzer.widgets.config_panel import ConfigPanel
 from driver_consistency_analyzer.widgets.chart_view import ChartView
 from driver_consistency_analyzer.widgets.corner_selector import CornerSelector
 
-__all__ = ["SessionPanel", "ConfigPanel", "ChartView", "CornerSelector"]
+__all__ = ["ConfigPanel", "ChartView", "CornerSelector"]
