+++ src/motorsports_data_notebook/desktop/__init__.py
@@ -1,0 +1,12 @@
+"""Shared desktop GUI components for motorsports data apps.
+
+Modules require tkinter and customtkinter at runtime. Import from
+submodules directly::
+
+    from motorsports_data_notebook.desktop.loader import load_session
+    from motorsports_data_notebook.desktop.session_panel import SessionPanel
+    from motorsports_data_notebook.desktop.dpi import get_screen_dpi, setup_hidpi_scaling
+    from motorsports_data_notebook.desktop.autocomplete_entry import AutocompleteEntry
+    from motorsports_data_notebook.desktop.stats_panel import BaseStatsPanel
+    from motorsports_data_notebook.desktop.config_panel import BaseConfigPanel
+"""
